add nav links for docs and leaderboard
