+++ src/lib/constants.ts
@@ -1,6 +1,7 @@
 export const TWITTER_URL = 'https://x.com/bamkfi'
 export const TELEGRAM_URL = 'https://t.me/bamkfi'
 export const GITHUB_URL = 'https://github.com/bamkfi/bamkfi-frontend'
+export const GITBOOK_URL = 'https://bamkfi.gitbook.io'
 export const BAMK_MARKET_URL = 'https://magiceden.io/runes/BAMK%E2%80%A2OF%E2%80%A2NAKAMOTO%E2%80%A2DOLLAR'
 export const NUSD_MARKET_URL = 'https://unisat.io/market/brc20?tick=%24NUSD'
 export const NUSD_RUNE_MARKET_URL = 'https://magiceden.io/runes/NUSD%E2%80%A2NUSD%E2%80%A2NUSD%E2%80%A2NUSD'
